+++ app/src/main/java/edu/ateneo/cie199/worky/FreelanceSignupActivity.java
@@ -76,7 +76,7 @@ public class FreelanceSignupActivity extends AppCompatActivity {
                     int fAge = Integer.parseInt(edtFage.getText().toString());
                     String fGender = LOOKUP_GENDER[spnFgender.getSelectedItemPosition()];
                     String fEmail = edtFemail.getText().toString();
-                    int fMobilenum = Integer.parseInt(edtFmobilenum.getText().toString());
+                    Long fMobilenum = Long.parseLong(edtFmobilenum.getText().toString());
                     String fProfile = edtFprofile.getText().toString();
                     String fEduc = edtFeduc.getText().toString();
                     String fExpertise = LOOKUP_FIELD[spnFexpertise.getSelectedItemPosition()];
